WIP revert UsersEditForm
